backend/requirements.txt
Django==5.0.14
djangorestframework==3.15.2
django-cors-headers==4.3.1
Pillow==10.4.0
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
dj-database-url==2.1.0
